% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_package_version}
\alias{check_package_version}
\title{Check if installed version of package is latest}
\usage{
check_package_version(package)
}
\arguments{
\item{package}{String. Package name.}
}
\value{
Side-effect. Raises a warning if package is outdated.
}
\description{
Check if installed version of package is latest
}
\details{
This function uses `pkgsearch::cran_package()` to check whether the installed version of a package is the latest available on CRAN. Useful to alert the user.
}
\examples{
\dontrun{
check_package_version("AER")
}
}
