% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{decode_wontrun}
\alias{decode_wontrun}
\title{Adds two columns to help decode the results of wontrun()}
\usage{
decode_wontrun(wontrun_df, ...)
}
\arguments{
\item{wontrun_df}{Data frame. A data frame as returned by `wontrun()`}
}
\value{
The input tibble with 2 more columns, `classes` and `cnd_message`
}
\description{
Adds two columns to help decode the results of wontrun()
}
\examples{
\dontrun{
aer_sources <- get_archived_sources("AER")
aer_runs <- aer_sources \%>\%
  wontrun(ncpus = 6, years = 2008)
summary_wontrun(aer_runs)
}
}
