% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_script_from_help}
\alias{generate_script_from_help}
\title{Generates an R script with the examples from an Rd file. Wraps `tools::Rd2ex()`}
\usage{
generate_script_from_help(path_to_rd)
}
\arguments{
\item{path_to_rd}{String. Path to Rd file you wish to convert to a script}
}
\value{
Side-effect. No returned object, writes a script on disk.
}
\description{
Generates an R script with the examples from an Rd file. Wraps `tools::Rd2ex()`
}
\details{
This wrapper around `tools::Rd2ex()` extracts the examples from Rd files included in
the man/ folder of source packages and writes them to disk, next to the Rd files.
}
\examples{
get_archived_sources("AER")
}
