% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_archived_sources}
\alias{get_archived_sources}
\title{Creates a tibble with the urls to the archived sources of a package}
\usage{
get_archived_sources(package)
}
\arguments{
\item{package}{String. Name of the package to query}
}
\value{
A tibble of 4 columns, `name`, `url`, `last_modified`, `size`
}
\description{
Creates a tibble with the urls to the archived sources of a package
}
\details{
The returned table is the same as the html table that can be found in the CRAN
archive url of a package, for instance \url{https://cran.r-project.org/src/contrib/Archive/AER/}.
The `url` column is added to make it easier to download a source package.
}
\examples{
get_archived_sources("AER")
}
