% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_available_packages}
\alias{get_available_packages}
\title{Get a tibble with current available packages on CRAN}
\usage{
get_available_packages(...)
}
\arguments{
\item{...}{Arguments passed down to `available.packages()`}
}
\value{
A tibble of 17 columns
}
\description{
Get a tibble with current available packages on CRAN
}
\details{
This function is a simple wrapper around `available.packages()`
}
\examples{
\dontrun{
get_available_packages()
}
}
