% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_example}
\alias{get_example}
\title{Downloads doc files from an archived source package, and generates an R script from it.}
\usage{
get_example(name, version, url, clean = TRUE, exdir = NULL)
}
\arguments{
\item{name}{String. Name of the packages.}

\item{version}{String. Version of the package to download.}

\item{url}{String. Link to archived package tar file.}

\item{clean}{Boolean, defaults to TRUE. If TRUE, only keeps man/ folder containing the documentaton. If FALSE, keeps entire package.}

\item{exdir}{String. Path to directory to save examples. Defaults to NULL, so a temporary directory gets used.}
}
\value{
Side-effect. No returned object, writes a Rd files to disk.
}
\description{
Downloads doc files from an archived source package, and generates an R script from it.
}
\details{
This function returns a data frame with a column `name` giving the name
of a package, `version` giving its version, `url` the download url
`last_modified` the date on which the package was last modified on CRAN
and `size`, the size of the package
}
