% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_packages_from_view}
\alias{get_packages_from_view}
\title{Get a tibble of packages from a CRAN Task View}
\usage{
get_packages_from_view(view, date = "2015-01-01")
}
\arguments{
\item{view}{String. View of interest}

\item{date}{String, format: '2015-01-01'. Queries MRAN to download state of the view at given date}
}
\value{
A tibble of 2 columns
}
\description{
Get a tibble of packages from a CRAN Task View
}
\details{
This function is a simple wrapper around `ctv::ctv()`.
}
\examples{
\dontrun{
get_packages_from_view("Econometrics")
get_packages_from_view("Bayesian", "2022-01-01")
}
}
