% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_sources_for_selected_packages}
\alias{get_sources_for_selected_packages}
\title{Get a tibble of source urls for a selection of packages}
\usage{
get_sources_for_selected_packages(view_df)
}
\arguments{
\item{view_df}{A data frame as returned by `get_packages_from_view()`}
}
\value{
A tibble of 5 columns.
}
\description{
Get a tibble of source urls for a selection of packages
}
\details{
This function returns a data frame with a column `name` giving the name
of a package, `version` giving its version, `url` the download url
`last_modified` the date on which the package was last modified on CRAN
and `size`, the size of the package
}
\examples{
\dontrun{
ctv_econ <- get_packages_from_view("Econometrics", date = "2015-01-01")
get_sources_for_selected_packages(ctv_econ)
}
}
