% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{n_packages_time}
\alias{n_packages_time}
\title{Number of packages in CRAN Task Views, per view}
\format{
A data frame with 515 rows and 3 columns
\describe{
  \item{view}{CRAN Task view.}
  \item{as_of}{A date. Number of packages as of this date.}
  \item{total_packages}{Total packages available in the view.}
}
}
\usage{
n_packages_time
}
\description{
Number of packages in CRAN Task Views, per view
}
\keyword{datasets}
