% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sources_ctv_econ}
\alias{sources_ctv_econ}
\title{A sample of packages from the 'Econometrics' view as of 2010.}
\format{
A data frame with 10 rows and 7 columns
\describe{
  \item{view}{CRAN Task view, only 'Econometrics' here.}
  \item{name}{A date. Number of packages as of this date.}
  \item{version}{Package version.}
  \item{url}{Download url.}
  \item{last_modified}{Date at which the package was last modified.}
  \item{size}{Size of the source tarball.}
}
}
\usage{
sources_ctv_econ
}
\description{
A sample of packages from the 'Econometrics' view as of 2010.
}
\keyword{datasets}
