% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summary_wontrun}
\alias{summary_wontrun}
\title{Summarise results from wontrun}
\usage{
summary_wontrun(wontrun_df)
}
\arguments{
\item{wontrun_df}{Data frame. A data frame as returned by `get_archived_sources()`}
}
\value{
Side-effect. No returned object, writes a Rd files to disk.
}
\description{
Summarise results from wontrun
}
\examples{
\dontrun{
aer_sources <- get_archived_sources("AER")
aer_runs <- aer_sources \%>\%
  wontrun(ncpus = 6, years = 2008)
summary_wontrun(aer_runs)
}
}
