% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{with_pload}
\alias{with_pload}
\title{Runs code with an attached package. Detaches it afterwards. For internal use only.}
\usage{
with_pload(package, code)
}
\arguments{
\item{package}{String. Package name. If the package is not available, it gets installed.}

\item{code}{Expression. Code to run in with this package attached.}
}
\description{
Runs code with an attached package. Detaches it afterwards. For internal use only.
}
