% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{wontrun}
\alias{wontrun}
\title{wontrun}
\usage{
wontrun(
  packages_df,
  ncpus = 1,
  years = NULL,
  earliest = TRUE,
  setup = FALSE,
  wontrun_lib
)
}
\arguments{
\item{packages_df}{Data frame. A data frame as returned by `get_archived_sources()`}

\item{ncpus}{Integer. Number of cpus to run examples in parallel.}

\item{years}{Integer. Year or atomic vector of years to select packages to run examples.}

\item{earliest}{Boolean. Select oldest package from year.}
}
\value{
Side-effect. No returned object, writes a Rd files to disk.
}
\description{
wontrun
}
\examples{
\dontrun{
aer_sources <- get_archived_sources("AER")
aer_runs <- aer_sources \%>\%
  wontrun(ncpus = 6, years = 2008)
}
}
