% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_script_from_help}
\alias{generate_script_from_help}
\title{Generates an R script with the examples from an Rd file.}
\usage{
generate_script_from_help(path_to_rd, rm_dont_run = TRUE)
}
\arguments{
\item{path_to_rd}{String. Path to Rd file you wish to convert to a script}

\item{rm_dont_run}{Boolean. Should scripts with \dontrun tag be ignored? Defaults to TRUE.}
}
\value{
Side-effect. No returned object, writes a script on disk.
}
\description{
Generates an R script with the examples from an Rd file.
}
\details{
This function is a wrapper around `tools::Rd2ex()` tht extracts the examples from Rd files included in
the man/ folder of source packages and writes them to disk, next to the Rd files.
}
\examples{
get_archived_sources("AER")
}
